tidy up of MediaApiController into suggestionController
