+++ media-api/app/controllers/SuggestionController.scala
@@ -1,0 +1,108 @@
+package controllers
+
+import play.api.libs.concurrent.Execution.Implicits._
+import play.api.libs.json.{Reads, Json, Writes}
+import play.api.mvc.{AnyContent, Request, Controller}
+
+import com.gu.mediaservice.lib.argo.ArgoHelpers
+import com.gu.mediaservice.lib.auth
+import com.gu.mediaservice.lib.auth.{PermissionStore, KeyStore}
+
+import lib.Config
+import lib.elasticsearch.{AggregateSearchResults, ElasticSearch}
+import lib.querysyntax.Parser
+
+import scala.util.Try
+
+object SuggestionController extends Controller with ArgoHelpers {
+
+  val keyStore = new KeyStore(Config.keyStoreBucket, Config.awsCredentials)
+  val permissionStore = new PermissionStore(Config.configBucket, Config.awsCredentials)
+
+  import Config.loginUriTemplate
+
+  val Authenticated = auth.Authenticated(keyStore, loginUriTemplate, Config.kahunaUri)
+
+  def suggestMetadataCredit(q: Option[String], size: Option[Int]) = Authenticated.async { request =>
+    ElasticSearch
+      .completionSuggestion("suggestMetadataCredit", q.getOrElse(""), size.getOrElse(10))
+      .map(c => respondCollection(c.results))
+  }
+
+  def suggestLabels(q: Option[String]) = Authenticated {
+
+    val pseudoFamousLabels = List(
+      "cities", "family", "filmandmusic", "lr", "pp", "saturdayreview", "travel",
+
+      "culturearts", "culturebooks", "culturefilm", "culturestage", "culutremusic",
+
+      "g2arts", "g2columns", "g2coverfeatures", "g2fashion", "g2features", "g2food", "g2health",
+      "g2lifestyle", "g2shortcuts", "g2tv", "g2women",
+
+      "obsbizcash", "obscomment", "obsfocus", "obsfoodfeat", "obsfoodother", "obsfoodrecipes",
+      "obsfoodsupp", "obsforeign", "obshome", "obsmagfash", "obsmagfeat", "obsmaglife",
+      "obsrevagenda", "obsrevbooks", "obsrevcritics", "obsrevdiscover", "obsrevfeat",
+      "obsrevmusic", "obsrevtv", "obssports", "obssupps", "obstechbright", "obstechfeat",
+      "obstechplay"
+    )
+
+    val labels = q.map { q =>
+      pseudoFamousLabels.filter(_.startsWith(q))
+    }.getOrElse(pseudoFamousLabels)
+
+    respondCollection(labels)
+  }
+
+  def suggestLabelSiblings(label: String, q: Option[String], selectedLabels: Option[String]) = Authenticated.async { request =>
+    val selectedLabels_ = selectedLabels.map(_.split(",").toList.map(_.trim)).getOrElse(Nil)
+    val structuredQuery = q.map(Parser.run) getOrElse List()
+
+    ElasticSearch.labelSiblingsSearch(structuredQuery, excludeLabels = List(label)) map { agg =>
+      val labels = agg.results.map { label =>
+        val selected = selectedLabels_.contains(label.key)
+        LabelSibling(label.key, selected)
+      }
+      respond(LabelSiblingsResponse(label, labels.toList))
+    }
+  }
+
+  // TODO: work with analysed fields
+  // TODO: recover with HTTP error if invalid field
+  // TODO: Add validation, especially if you use length
+  def metadataSearch(field: String) = Authenticated.async { request =>
+    ElasticSearch.metadataSearch(AggregateSearchParams(field, request)) map aggregateResponse
+  }
+
+  def editsSearch(field: String) = Authenticated.async { request =>
+    ElasticSearch.editsSearch(AggregateSearchParams(field, request)) map aggregateResponse
+  }
+
+  // TODO: Add some useful links
+  def aggregateResponse(agg: AggregateSearchResults) =
+    respondCollection(agg.results, Some(0), Some(agg.total))
+
+}
+
+case class LabelSibling(name: String, selected: Boolean)
+object LabelSibling {
+  implicit def jsonWrites: Writes[LabelSibling] = Json.writes[LabelSibling]
+  implicit def jsonReads: Reads[LabelSibling] =  Json.reads[LabelSibling]
+}
+
+case class LabelSiblingsResponse(label: String, siblings: List[LabelSibling])
+object LabelSiblingsResponse {
+  implicit def jsonWrites: Writes[LabelSiblingsResponse] = Json.writes[LabelSiblingsResponse]
+  implicit def jsonReads: Reads[LabelSiblingsResponse] =  Json.reads[LabelSiblingsResponse]
+}
+
+case class AggregateSearchParams(field: String, q: Option[String])
+object AggregateSearchParams {
+  def parseIntFromQuery(s: String): Option[Int] = Try(s.toInt).toOption
+
+  def apply(field: String, request: Request[AnyContent]): AggregateSearchParams = {
+    AggregateSearchParams(
+      field,
+      request.getQueryString("q")
+    )
+  }
+}
