+++ media-api/app/lib/elasticsearch/ElasticSearch.scala
@@ -161,17 +161,17 @@ object ElasticSearch extends ElasticSearchClient with SearchFilters with ImageFi
   }
 
   def metadataSearch(params: AggregateSearchParams)(implicit ex: ExecutionContext): Future[AggregateSearchResults] =
-    aggregateSearch("metadata", metadataField(params.field), params.q)
+    aggregateSearch("metadata", metadataField(params.field), params)
 
   def editsSearch(params: AggregateSearchParams)(implicit ex: ExecutionContext): Future[AggregateSearchResults] =
-    aggregateSearch("edits", editsField(params.field), params.q)
+    aggregateSearch("edits", editsField(params.field), params)
 
-  def aggregateSearch(name: String, field: String, q: Option[String])
+  def aggregateSearch(name: String, field: String, params: AggregateSearchParams)
                      (implicit ex: ExecutionContext): Future[AggregateSearchResults] = {
     val aggregate = AggregationBuilders
       .terms(name)
       .field(field)
-      .include(Pattern.quote(q.getOrElse("")) + ".*", Pattern.CASE_INSENSITIVE)
+      .include(Pattern.quote(params.q.getOrElse("")) + ".*", Pattern.CASE_INSENSITIVE)
 
     val search = prepareImagesSearch.addAggregation(aggregate)
